o Bugfix: premature victory in combat in slime cave. Npc's were not being added   properly to the combat npc array when they were introduced after start of   combat. The array elements were getting overwritten and the max used index   was not updated right. Fixed additions and got rid of the max-used index.
o Bugfix: hitting the '<' key in dungeons caused a crash. Disabled this command
  in dungeons.

+++ src/common.h
@@ -324,7 +324,7 @@ extern "C" {
 #define max(a,b) ((a) > (b) ? (a) : (b))
 #define min(a,b) ((a) < (b) ? (a) : (b))
 #define clamp(v,a,b) ((v) = (v) < (a) ? (a) : ((v) > (b) ? (b) : (v)))
-#define array_sz(a) (sizeof((a))/sizeof((a)[0]))
+#define array_sz(a) (int)(sizeof((a))/sizeof((a)[0]))
 #define outcast(ptr,type,field) \
         ((type*)((char*)(ptr)-(unsigned long)(&((type *)0)->field)))
 #define CREATE(ptr,type,err) \
